fix: renames shouldShowSupporterPlusMessaging to isSupporterPlus and moves it alongside similar checks
