+++ support-frontend/assets/helpers/redux/checkout/product/selectors/isSupporterPlus.ts
@@ -1,8 +1,35 @@
+import type {
+	ContributionType,
+	OtherAmounts,
+	SelectedAmounts,
+} from 'helpers/contributions';
 import { getAmount } from 'helpers/contributions';
+import type { CountryGroupId } from 'helpers/internationalisation/countryGroup';
 import { getContributionType } from 'helpers/redux/checkout/product/selectors/productType';
 import type { ContributionsState } from 'helpers/redux/contributionsStore';
 import { getThresholdPrice } from 'helpers/supporterPlus/benefitsThreshold';
 import { isOneOff } from 'helpers/supporterPlus/isContributionRecurring';
+
+export function isSupporterPlus(
+	contributionType: ContributionType,
+	selectedAmounts: SelectedAmounts,
+	otherAmounts: OtherAmounts,
+	countryGroupId: CountryGroupId,
+): boolean {
+	if (isOneOff(contributionType)) {
+		return false;
+	}
+
+	const benefitsThreshold = getThresholdPrice(countryGroupId, contributionType);
+	const selectedAmount = selectedAmounts[contributionType];
+
+	if (selectedAmount === 'other') {
+		const otherAmount = otherAmounts[contributionType].amount;
+		return otherAmount ? parseInt(otherAmount) >= benefitsThreshold : false;
+	}
+
+	return selectedAmount >= benefitsThreshold;
+}
 
 export function isSupporterPlusPurchase(state: ContributionsState): boolean {
 	const contributionType = getContributionType(state);
